تست مرحله نهایی دانش آموز و معلم

+++ src/Services/Main/Main.Application/Contracts/Persistence/IStudentRepository.cs
@@ -4,6 +4,6 @@ namespace Main.Application.Contracts.Persistence
 {
     public interface IStudentRepository : IGenericRepository<Student>
     {
-        Task InActiveAsync(int id);
+        Task InActiveAsync(long id);
     }
 }
