+++ src/Services/Main/Main.Application/Contracts/Persistence/ITeacherRepository.cs
@@ -4,6 +4,6 @@ namespace Main.Application.Contracts.Persistence
 {
     public interface ITeacherRepository : IGenericRepository<Teacher>
     {
-        Task InActiveAsync(int id);
+        Task InActiveAsync(long id);
     }
 }
