+++ src/Services/Main/Main.Application/Features/Students/Commands/DeleteStudent/DeleteStudentCommand.cs
@@ -3,7 +3,7 @@ using SharedLibrary.Patterns.ResultPattern;
 
 namespace Main.Application.Features.Students.Commands.DeleteStudent
 {
-    public record DeleteStudentCommand(int Id) : IRequest<Result<bool>>
+    public record DeleteStudentCommand(long Id) : IRequest<Result<bool>>
     {
 
     }
