+++ src/Services/Main/Main.Domain/Models/StudentCourse.cs
@@ -4,10 +4,10 @@ namespace Main.Domain.Models
 {
     public class StudentCourse : EntityBase
     {
-        public int StudentId { get; set; }
+        public long StudentId { get; set; }
         public Student Student { get; set; } = null!;
 
-        public int CourseId { get; set; }
+        public long CourseId { get; set; }
         public Course Course { get; set; } = null!;
     }
 }
