+++ src/Services/Main/Main.Infrastructure/Repositories/TeacherRepository.cs
@@ -12,7 +12,7 @@ namespace Main.Infrastructure.Repositories
             _context = context;
         }
 
-        public async Task InActiveAsync(int id)
+        public async Task InActiveAsync(long id)
         {
             var teacher = _context.Teachers.FirstOrDefault(x => x.Id == id);
             if (teacher == null)
